+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Target `angular`: Add `build:<target>:assets`, `build:<target>:pug` and `build:<target>:static`.
 - Update `gulp-typescript`: solve error message during compilation
 - Targets `node` and `angular`: `build:<target>:scripts` now include in-lined source maps
+- Target `node`: `watch:<target>` to support incremental builds